Move file download buttons into own module.
